+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/utils/IdeServicesPluginRepositoryService.kt
@@ -23,14 +23,14 @@ interface IdeServicesPluginRepositoryService : PluginRepositoryService {
         @Part file: MultipartBody.Part,
     ): Call<PluginUpdateBean>
 
-    @Deprecated("Use uploadByStringIdAndFamily(pluginXmlId, family, channel, notes, isHidden, file)")
     @Multipart
     @POST("/api/plugins")
-    override fun uploadByStringId(
-        @Part("xmlId") pluginXmlId: RequestBody,
-        @Part("channel") channel: RequestBody?,
-        @Part("notes") notes: RequestBody?,
-        @Part("isHidden") isHidden: Boolean,
-        @Part file: MultipartBody.Part,
+    override fun uploadByStringIdAndFamily(
+        pluginXmlId: RequestBody,
+        family: RequestBody,
+        channel: RequestBody?,
+        notes: RequestBody?,
+        isHidden: Boolean,
+        file: MultipartBody.Part,
     ): Call<PluginUpdateBean>
 }